otioload: Add simple dialog to change settings
Adds a simple table of sequences in the otio file and allows the
settings to be changed. The dialog takes a list of sequences and relies
on each otio timeline having a unique name.

The dialog has to be called from the main thread so there is a wrapper
function in Core.